Added second page of etudes covering more techniques
Three studies were added. One to teach and focus on the clarion register of the instrument, slurred and smooth. The second focuses on working back and forth over the break using the skills of right hand down and pushing the air. The final etude aims to work as an intruduction to correct articulation technique using repetitive notes in the throat tones.