B-flat major with study tips and etude duet
Top of the page has a schedule of how to practice, how long to practice each thing, and what order to do it in. It is at the top of the page so the student will see it first thing.

Etude for this key is a duet. Focus is on the important fingerings for B-flat major. Side E-flat in low register and 1-and-1 in the upper, and using the left pinky when going from E-flat to C-natural in the clarion register.

Teacher part is in triplets so the student can learn to hear in the subdivision of three without having to play in it. This will be a first step into playing in 6/8 time.